etl redcap-det kiosk: Convert string age_months to float, not int
int() won't work when the string represents a float.

This was a bug on my part when I was applying age_ceiling().

+++ lib/seattleflu/id3c/cli/command/etl/redcap_det_kiosk.py
@@ -800,7 +800,7 @@ def determine_all_questionnaire_items(redcap_record: dict) -> List[dict]:
 
     if redcap_record['age']:
         items['age'] = [{ 'valueInteger': age_ceiling(int(redcap_record['age'])) }]
-        items['age_months'] = [{ 'valueInteger': int(age_ceiling(int(redcap_record['age_months']) / 12) * 12) }]
+        items['age_months'] = [{ 'valueInteger': int(age_ceiling(float(redcap_record['age_months']) / 12) * 12) }]
 
     # Participant can select multiple insurance types, so create
     # a separate answer for each selection
